+++ src/components/modules/contact/hooks/contact.hook.ts
@@ -12,25 +12,7 @@ import {
 } from "firebase/firestore";
 import { useToast } from "@/hooks/use-toast";
 import { db } from "@/core/config/firebase/firebase";
-
-export enum WalletType {
-  Albedo = "Albedo",
-  LOBSTR = "LOBSTR",
-}
-
-const formSchema = z.object({
-  name: z.string().min(3, {
-    message: "Name must be at least 3 characters.",
-  }),
-  lastName: z.string().min(3, {
-    message: "Last name must be at least 3 characters.",
-  }),
-  email: z.string().email({ message: "Invalid email address" }),
-  type: z.nativeEnum(WalletType),
-  address: z.string().min(10, {
-    message: "Address must be at least 10 characters.",
-  }),
-});
+import { formSchema } from "../schema/contact-schema";
 
 export const useContact = () => {
   const { toast } = useToast();
